+++ packages/svelte/CHANGELOG.md
@@ -6,6 +6,8 @@ description: All notable changes will be documented in this file.
 
 ## [Unreleased]
 
+## [5.19.0] - 2026-02-27
+
 ### Added
 
 - **Format**: Added `Format.Time` for formatting time values and examples for basic, date input, seconds, custom AM/PM